+++ migrations.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    'migrations_paths' => [
+        'Kuusamo\Vle\Migration' => __DIR__ . '/src/Migration'
+    ],
+    'all_or_nothing' => true,
+    'check_database_platform' => true
+];