+++ test/Entity/FileTest.php
@@ -9,6 +9,8 @@ class FileTest extends TestCase
 {
     public function testAccessors()
     {
+        $folderMock = $this->createMock('Kuusamo\Vle\Entity\Folder');
+
         $file = new File;
 
         $file->setId(10);
@@ -16,12 +18,14 @@ class FileTest extends TestCase
         $file->setFilename('document.pdf');
         $file->setMediaType('application/pdf');
         $file->setSize(123456);
+        $file->setFolder($folderMock);
 
-        $this->assertEquals(10, $file->getId());
-        $this->assertEquals('PDF', $file->getName());
-        $this->assertEquals('document.pdf', $file->getFilename());
-        $this->assertEquals('application/pdf', $file->getMediaType());
-        $this->assertEquals(123456, $file->getSize());
+        $this->assertSame(10, $file->getId());
+        $this->assertSame('PDF', $file->getName());
+        $this->assertSame('document.pdf', $file->getFilename());
+        $this->assertSame('application/pdf', $file->getMediaType());
+        $this->assertSame(123456, $file->getSize());
+        $this->assertSame($folderMock, $file->getFolder());
 
         $this->assertSame(
             '{"id":10,"name":"PDF","filename":"document.pdf"}',